Move extract children from rules to utils
